+++ src/se/osbe/emoreactor/brain/emotions/AfraidEmotion.java
@@ -3,7 +3,7 @@ package se.osbe.emoreactor.brain.emotions;
 import se.osbe.emoreactor.brain.reactor.ReactorException;
 
 public class AfraidEmotion extends AbstractEmotion {
-	public AfraidEmotion(Double amplitude, long initialTime, long duration) throws ReactorException {
+	public AfraidEmotion(Float amplitude, long initialTime, long duration) throws ReactorException {
 		super(EmotionType.AFRAID, amplitude, initialTime, duration);
 	}
 }
