+++ src/se/osbe/emoreactor/brain/emotions/AngerEmotion.java
@@ -3,7 +3,7 @@ package se.osbe.emoreactor.brain.emotions;
 import se.osbe.emoreactor.brain.reactor.ReactorException;
 
 public class AngerEmotion extends AbstractEmotion {
-	public AngerEmotion(Double amplitude, long initialTime, long duration) throws ReactorException {
+	public AngerEmotion(Float amplitude, long initialTime, long duration) throws ReactorException {
 		super(EmotionType.ANGER, amplitude, initialTime, duration);
 	}
 }
