+++ src/se/osbe/emoreactor/brain/emotions/InterestedEmotion.java
@@ -4,7 +4,7 @@ import se.osbe.emoreactor.brain.reactor.ReactorException;
 
 public class InterestedEmotion extends AbstractEmotion {
 	
-	public InterestedEmotion(Double amplitude, long initialTime, long duration) throws ReactorException {
+	public InterestedEmotion(Float amplitude, long initialTime, long duration) throws ReactorException {
 		super(EmotionType.INTERESTED, amplitude, initialTime, duration);
 	}
 }
