+++ src/se/osbe/emoreactor/brain/emotions/JudgementalEmotion.java
@@ -4,7 +4,7 @@ import se.osbe.emoreactor.brain.reactor.ReactorException;
 
 public class JudgementalEmotion extends AbstractEmotion {
 	
-	public JudgementalEmotion(Double amplitude, long initialTime, long duration) throws ReactorException {
+	public JudgementalEmotion(Float amplitude, long initialTime, long duration) throws ReactorException {
 		super(EmotionType.JUDGEMENTAL, amplitude, initialTime, duration);
 	}
 }
