+++ src/se/osbe/emoreactor/brain/emotions/OpenEmotion.java
@@ -3,7 +3,7 @@ package se.osbe.emoreactor.brain.emotions;
 import se.osbe.emoreactor.brain.reactor.ReactorException;
 
 public class OpenEmotion extends AbstractEmotion {
-	public OpenEmotion(Double amplitude, long initialTime, long duration) throws ReactorException {
+	public OpenEmotion(Float amplitude, long initialTime, long duration) throws ReactorException {
 		super(EmotionType.OPEN, amplitude, initialTime, duration);
 	}
 }
