+++ src/se/osbe/emoreactor/brain/emotions/PeacefulEmotion.java
@@ -3,7 +3,7 @@ package se.osbe.emoreactor.brain.emotions;
 import se.osbe.emoreactor.brain.reactor.ReactorException;
 
 public class PeacefulEmotion extends AbstractEmotion {
-	public PeacefulEmotion(Double amplitude, long initialTime, long duration) throws ReactorException {
+	public PeacefulEmotion(Float amplitude, long initialTime, long duration) throws ReactorException {
 		super(EmotionType.PEACEFUL, amplitude, initialTime, duration);
 	}
 }
