+++ src/se/osbe/emoreactor/brain/emotions/ReliefEmotion.java
@@ -3,7 +3,7 @@ package se.osbe.emoreactor.brain.emotions;
 import se.osbe.emoreactor.brain.reactor.ReactorException;
 
 public class ReliefEmotion extends AbstractEmotion {
-	public ReliefEmotion(Double amplitude, long initialTime, long duration) throws ReactorException {
+	public ReliefEmotion(Float amplitude, long initialTime, long duration) throws ReactorException {
 		super(EmotionType.RELIEF, amplitude, initialTime, duration);
 	}
 }
