+++ src/se/osbe/emoreactor/brain/emotions/SadEmotion.java
@@ -4,7 +4,7 @@ import se.osbe.emoreactor.brain.reactor.ReactorException;
 
 public class SadEmotion extends AbstractEmotion {
 	
-	public SadEmotion(Double amplitude, long initialTime, long duration) throws ReactorException {
+	public SadEmotion(Float amplitude, long initialTime, long duration) throws ReactorException {
 		super(EmotionType.SADNESS, amplitude, initialTime, duration);
 	}
 }
