+++ src/se/osbe/emoreactor/brain/emotions/StrongEmotion.java
@@ -4,7 +4,7 @@ import se.osbe.emoreactor.brain.reactor.ReactorException;
 
 public class StrongEmotion extends AbstractEmotion {
 	
-	public StrongEmotion(Double amplitude, long initialTime, long duration) throws ReactorException {
+	public StrongEmotion(Float amplitude, long initialTime, long duration) throws ReactorException {
 		super(EmotionType.STRONG, amplitude, initialTime, duration);
 	}
 }
